+++ app/src/main/java/com/freshdigitable/yttt/compose/AppSettingsScreen.kt
@@ -18,17 +18,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.freshdigitable.yttt.data.source.local.AndroidPreferencesDataStore
+import com.freshdigitable.yttt.AppSettingsViewModel
 import com.google.accompanist.themeadapter.material.MdcTheme
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-import javax.inject.Inject
 
 @Composable
 fun AppSettingsScreen(
@@ -79,25 +70,5 @@ private fun AppSettingsScreen(
 fun AppSettingsScreenPreview() {
     MdcTheme {
         AppSettingsScreen(text = { "24:00" }, onClicked = {})
-    }
-}
-
-@HiltViewModel
-class AppSettingsViewModel @Inject constructor(
-    private val dataStore: AndroidPreferencesDataStore,
-) : ViewModel() {
-    val changeDateTime: StateFlow<String> = dataStore.changeDateTime
-        .map { it ?: 24 }
-        .map { "$it:00" }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, "24:00")
-
-    fun onClick(value: Int) {
-        viewModelScope.launch {
-            dataStore.putTimeToChangeDate(value)
-        }
-    }
-
-    companion object {
-        val changeDateTimeList: IntRange = 24..27
     }
 }
